fix(#1701): roborev R9-F1 — a targeted read must not report the whole table
Under `--features tombstones` the partition-TARGETED materializing scans are
served by scanning the table and filtering (`retain`) to the partition key. The
inner full `scan` started its own `ReadOpMeter` and recorded the FULL result at
both its exits, so a `WHERE pk = ?` reported every row and every partition of
the table: the existing `a_cross_generation_point_read_is_one_metered_operation`
case measured `read.rows = 8.0` for a read that delivered 1 row from 1 of 8
partitions. That contradicts the "rows a read DELIVERED" contract catalog.rs
states, fails #1701's honesty AC in that build, and made one metric mean two
things across builds (the `not(tombstones)` path meters the delivered partition).

ONE meter per logical read, owned by the OUTER targeted API:
- `scan` / `scan_with_cell_metadata` become thin wrappers over new private
  `*_with_meter` inners (the R5-F4 already-started-meter mechanism), passing
  `ReadOpMeter::start(None)` — public signatures byte-identical, so no API break.
- the two `tombstones` fallbacks pass `ReadOpMeter::inert()` (construction-time
  no-op, distinct from the runtime `discard()`) and meter their own rows AFTER
  the retain.

Paid for the wrapper lines by moving both fallbacks into a sibling module
(precedent: manager_point_read.rs, reverse_scan.rs): mod.rs 3093 -> 3089 lines,
still under its 3138-line baseline, no CQLITE_ALLOW_FILE_GROWTH.
